refactor: reorg concrete/lib dir
